Create unit testing infrastructure, add first test
A testing suite has been created for the executive_smach_tutorials
submodule that allows for SMACH code to be generated from SMACHAML
scripts (SMACHAML: "SMACHA Ain't Markup Language") and SMACHA templates,
the generated code to be diff compared with the original code from the
tutorials, and the result to be unit-tested.

The first such unit test has been added for the 'state_machine2.py'
SMACH tutorial script.