Significant changes to attempt to fix state header nesting
Add ability to specify buffer type as dict or list. A buffer type of
dict allows for rendering the buffer to a template dict variable that
can be indexed by container state name, thus allowing container
templates to selectively render data that is only relevant to them, e.g.
userdata for that specific container.