Add major part of new 'contain' tool
The new 'contain' tool aims to be able to automatically wrap a sequence
of states in a script in a container state.

This commit adds a working, yet incomplete, prototype of the tool. State
transitions and paramaters now appear to be handled appropriately, but
userdata and remappings have yet to be accounted for.